refactor(mobile): extract document data layer from DocumentsScreen
Split the pure data layer (Document/ApiDocument types + the tested
subfoldersAtPath / folderNodeToDocument / pickDocumentType / toUiDocument
mappers) out of the 829-line DocumentsScreen.tsx into a sibling
documentModel.ts. DocumentsScreen re-exports the same public surface, so
the index barrel, sibling sheets, and unit tests import unchanged.

Behavior-preserving: no logic touched, only moved. Isolates data-layer
churn (e.g. the PR #918 field-mapping fixes) from presentation churn to
shrink the merge-conflict surface on this churn hotspot.
